% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_raw_enrollment.R
\name{aggregate_to_district}
\alias{aggregate_to_district}
\title{Aggregate school data to district level}
\usage{
aggregate_to_district(school_df)
}
\arguments{
\item{school_df}{School-level data frame}
}
\value{
District-level data frame
}
\description{
Aggregate school data to district level
}
\keyword{internal}
