% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_raw_enrollment.R
\name{build_tdoe_url}
\alias{build_tdoe_url}
\title{Build URL for TDOE data download}
\usage{
build_tdoe_url(end_year, file_type)
}
\arguments{
\item{end_year}{School year end}

\item{file_type}{Type of file ("membership", "profile", etc.)}
}
\value{
URL string
}
\description{
Constructs the URL for downloading data from Tennessee DOE.
}
\keyword{internal}
