% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_raw_assessment.R
\name{create_empty_assessment_raw}
\alias{create_empty_assessment_raw}
\title{Create empty assessment raw data frame}
\usage{
create_empty_assessment_raw()
}
\value{
Empty data frame with assessment columns
}
\description{
Returns an empty data frame with expected column structure for assessment data.
}
\keyword{internal}
