% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_assessment.R
\name{create_empty_assessment_result}
\alias{create_empty_assessment_result}
\title{Create empty assessment result data frame}
\usage{
create_empty_assessment_result(end_year)
}
\arguments{
\item{end_year}{School year end}
}
\value{
Empty data frame with expected columns
}
\description{
Create empty assessment result data frame
}
\keyword{internal}
