% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_assessment.R
\name{create_empty_tidy_assessment}
\alias{create_empty_tidy_assessment}
\title{Create empty tidy assessment data frame}
\usage{
create_empty_tidy_assessment()
}
\value{
Empty data frame with tidy assessment columns
}
\description{
Create empty tidy assessment data frame
}
\keyword{internal}
