% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_raw_assessment.R
\name{download_assessment_file}
\alias{download_assessment_file}
\title{Download a single assessment file}
\usage{
download_assessment_file(end_year, level)
}
\arguments{
\item{end_year}{School year end}

\item{level}{One of "state", "district", "school"}
}
\value{
Data frame or empty data frame if download fails
}
\description{
Download a single assessment file
}
\keyword{internal}
