% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_raw_enrollment.R
\name{download_tdoe_file}
\alias{download_tdoe_file}
\title{Download TDOE file}
\usage{
download_tdoe_file(url, file_type)
}
\arguments{
\item{url}{URL to download}

\item{file_type}{Type of file (for temp file naming)}
}
\value{
Data frame or NULL if download fails
}
\description{
Downloads a file from Tennessee DOE and reads it as Excel.
}
\keyword{internal}
