% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{get_assessment_cache_path}
\alias{get_assessment_cache_path}
\title{Get assessment cache file path for given year and type}
\usage{
get_assessment_cache_path(end_year, type, level = "all")
}
\arguments{
\item{end_year}{School year end}

\item{type}{Data type ("tidy" or "wide")}

\item{level}{Optional level filter that was used ("all", "state", "district", "school")}
}
\value{
Full path to cache file
}
\description{
Get assessment cache file path for given year and type
}
\keyword{internal}
