% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_raw_assessment.R
\name{get_raw_assessment}
\alias{get_raw_assessment}
\title{Download raw assessment data from TDOE}
\usage{
get_raw_assessment(end_year, level = "all")
}
\arguments{
\item{end_year}{School year end (2019, 2021-2025; no 2020 due to COVID waiver)}

\item{level}{One of "all", "state", "district", "school"}
}
\value{
List with state, district, and/or school data frames (depending on level)
}
\description{
Downloads state, district, and/or school assessment data from Tennessee DOE's
accountability portal.
}
\keyword{internal}
