% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_raw_enrollment.R
\name{get_raw_enr}
\alias{get_raw_enr}
\title{Download raw enrollment data from TDOE}
\usage{
get_raw_enr(end_year)
}
\arguments{
\item{end_year}{School year end (2023-24 = 2024)}
}
\value{
List with school and district data frames
}
\description{
Downloads school and district enrollment data from Tennessee DOE's data portal.
Uses Excel files from the Data Downloads page.
}
\keyword{internal}
