% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{make_campus_id}
\alias{make_campus_id}
\title{Create full campus ID from district and school IDs}
\usage{
make_campus_id(district_id, school_id)
}
\arguments{
\item{district_id}{District ID (will be padded to 4 digits)}

\item{school_id}{School ID (will be padded to 4 digits)}
}
\value{
Character vector of 8-digit campus IDs
}
\description{
Tennessee campus IDs are 8 digits: 4-digit district ID + 4-digit school ID
}
\keyword{internal}
