% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{pad_district_id}
\alias{pad_district_id}
\title{Pad district ID to 4 digits}
\usage{
pad_district_id(x)
}
\arguments{
\item{x}{Vector of district IDs}
}
\value{
Character vector of zero-padded district IDs
}
\description{
Tennessee district IDs are 4 digits, zero-padded
}
\keyword{internal}
