% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{pad_school_id}
\alias{pad_school_id}
\title{Pad school ID to 4 digits}
\usage{
pad_school_id(x)
}
\arguments{
\item{x}{Vector of school IDs}
}
\value{
Character vector of zero-padded school IDs
}
\description{
Tennessee school IDs within a district are 4 digits, zero-padded
}
\keyword{internal}
