% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_assessment.R
\name{process_assessment_level}
\alias{process_assessment_level}
\title{Process a single level of assessment data}
\usage{
process_assessment_level(df, end_year, type)
}
\arguments{
\item{df}{Raw data frame for one level (state/district/school)}

\item{end_year}{School year end}

\item{type}{Record type ("State", "District", "School")}
}
\value{
Processed data frame
}
\description{
Process a single level of assessment data
}
\keyword{internal}
