% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{read_assessment_cache}
\alias{read_assessment_cache}
\title{Read assessment data from cache}
\usage{
read_assessment_cache(end_year, type, level = "all")
}
\arguments{
\item{end_year}{School year end}

\item{type}{Data type ("tidy" or "wide")}

\item{level}{Optional level filter that was used ("all", "state", "district", "school")}
}
\value{
Cached data frame
}
\description{
Read assessment data from cache
}
\keyword{internal}
