% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_assessment.R
\name{standardize_subgroup}
\alias{standardize_subgroup}
\title{Standardize subgroup names}
\usage{
standardize_subgroup(x)
}
\arguments{
\item{x}{Vector of subgroup names}
}
\value{
Standardized subgroup names
}
\description{
Standardize subgroup names
}
\keyword{internal}
