% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_assessment.R
\name{standardize_subject}
\alias{standardize_subject}
\title{Standardize subject names}
\usage{
standardize_subject(x)
}
\arguments{
\item{x}{Vector of subject names}
}
\value{
Standardized subject names
}
\description{
Standardize subject names
}
\keyword{internal}
